refactor: prettyjson
