+++ internal/prettyjson/marshal.go
@@ -1,10 +1,11 @@
-package utils
+package prettyjson
 
 import (
 	"encoding/json"
 )
 
-func PrettyJSON(input interface{}) (string, error) {
+// Marshals structs into stringified JSON with indenting.
+func Marshal(input any) (string, error) {
 	output, err := json.MarshalIndent(input, "", "    ")
 	if err != nil {
 		return "", err
